♻️ refactor : 모달 로직 위치 변경 및 기존 스크롤 형태 페이지 삭제 #55
